% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{`explain` the reason for an error message}
\usage{
explain(x)
}
\arguments{
\item{x}{An error message, for example
\code{x = "Error in c[1] : object of type 'builtin' is not subsettable"}}
}
\description{
`explain` the reason for an error message
}
\examples{
explain("Some random error message")
explain("Error in c[1] : object of type 'builtin' is not subsettable")

}
