% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_subsettable.R
\docType{data}
\name{not_subsettable}
\alias{not_subsettable}
\title{Error: "object of type <XXX> is not subsettable"}
\format{An object of class \code{NULL} of length 0.}
\usage{
not_subsettable
}
\description{
Blah blah blah
}
\keyword{datasets}
