% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{wtf}
\alias{wtf}
\title{"WTF" does this error message mean?}
\usage{
wtf(x)
}
\arguments{
\item{x}{An error message, for example
\code{x = "Error in c[1] : object of type 'builtin' is not subsettable"}}
}
\description{
"WTF" does this error message mean?
}
